// chapter-4/api/src/modules/cars/dtos/iCreateCar.dto.ts
import { Specification } from "../infra/typeorm/entities/specification";

interface ICreateCarDTO {
  id?: string,
  name: string,
  description: string;
  daily_rate: number;
  license_plate: string;
  fine_amount: number;
  brand: string;
  category_id: string;
  specifications?: Specification[];
}

export { ICreateCarDTO };